requirements.txt
streamlit
Pillow
ollama
python-docx
PyPDF2
pyttsx3
SpeechRecognition
pandas
numpy
altair
joblib
scikit-learn
